Restructure AI Glasses template into androidstudio activity hierarchy
