+++ utilities/templates-impl/src/main/java/com/itsaky/androidide/templates/impl/androidstudio/activity/aiGlassesActivity/aiGlassesActivityManifest.kt
@@ -1,0 +1,5 @@
+package com.itsaky.androidide.templates.impl.androidstudio.activity.aiGlassesActivity
+
+// Kept for Android Studio template parity. Manifest generation is implemented via templates-api
+// manifest DSL in aiGlassesActivityRecipe.kt.
+internal const val aiGlassesActivityManifestMarker = "aiGlassesActivityManifest"
