+++ utilities/templates-impl/src/main/java/com/itsaky/androidide/templates/impl/androidstudio/activity/aiGlassesActivity/res/values/stringsXml.kt
@@ -1,0 +1,16 @@
+package com.itsaky.androidide.templates.impl.androidstudio.activity.aiGlassesActivity.res.values
+
+internal fun stringsXml() =
+    """
+<resources>
+    <string name="app_name">Basic AI Glasses Activity</string>
+    <string name="display_category_xr_projected">com.google.android.glasses.category.PROJECTED</string>
+    <string name="hello_ai_glasses">Hello, AI Glasses!</string>
+    <string name="launch">Launch</string>
+    <string name="status_prefix">Status: </string>
+    <string name="status_connected">Connected</string>
+    <string name="status_disconnected">Disconnected</string>
+    <string name="unsupported_android_version">This version of Android is not supported</string>
+    <string name="close">Close</string>
+</resources>
+""".trim()
