+++ utilities/templates-impl/src/main/java/com/itsaky/androidide/templates/impl/androidstudio/activity/aiGlassesActivity/src/app_package/audioInterfaceKt.kt
@@ -1,0 +1,40 @@
+package com.itsaky.androidide.templates.impl.androidstudio.activity.aiGlassesActivity.src.app_package
+
+import com.itsaky.androidide.templates.base.AndroidModuleTemplateBuilder
+
+internal fun AndroidModuleTemplateBuilder.audioInterfaceKt() =
+    """
+package ${data.packageName}
+
+import android.content.Context
+import android.speech.tts.TextToSpeech
+import android.util.Log
+import androidx.lifecycle.DefaultLifecycleObserver
+import androidx.lifecycle.LifecycleOwner
+
+class AudioInterface(private val context: Context, private val initializationMessage: String) :
+    DefaultLifecycleObserver {
+  private lateinit var tts: TextToSpeech
+
+  override fun onStart(owner: LifecycleOwner) {
+    super.onStart(owner)
+    tts = TextToSpeech(context) { status ->
+      if (status == TextToSpeech.SUCCESS) speak(initializationMessage)
+      else Log.e(TAG, "Initialization failed with status: $status")
+    }
+  }
+
+  fun speak(textToSpeak: String) {
+    tts.speak(textToSpeak, TextToSpeech.QUEUE_ADD, null, initializationMessage.lowercase().replace(" ", "_"))
+  }
+
+  override fun onStop(owner: LifecycleOwner) {
+    super.onStop(owner)
+    tts.shutdown()
+  }
+
+  companion object {
+    private const val TAG = "AudioInterface"
+  }
+}
+""".trim()
